requirements/requirements.txt
fastapi[all]
tortoise-orm[asyncpg]
aerich